Yeah so completely reworked the QML there.
MainTabbedPane opens a Navigation, which has a ListPage as a child.
Triggering a task list item will give you the EditPage, whereas
triggering anything else will give you a ListPageChild, with the
datamodel filter updated to only include items that are associated with
the triggered item (which will only be tasks, which, when triggered, do
the EditPage again). yo siiiiiick.